add hidden v2 portfolio route scaffold
