+++ src/App.jsx
@@ -1,63 +1,10 @@
-import { Header, About, Projects } from "./containers";
-import { Navbar, CTA, Contact } from "./components";
-import React, {useEffect, useRef} from "react";
-import Skills from "./containers/skills/Skills";
-import "./App.css";
+import PortfolioCurrent from "./pages/PortfolioCurrent";
+import PortfolioV2 from "./pages/PortfolioV2";
 
 function App() {
-  const sectionsRef = useRef([]);
+  const route = window.location.hash;
 
-  useEffect(() => {
-    const observer = new IntersectionObserver(
-      (entries) => {
-        entries.forEach((entry) => {
-          if (entry.isIntersecting) {
-            entry.target.classList.add("visible")
-          } else {
-            entry.target.classList.remove("visible")
-          }
-        });
-      },
-      {threshold: 0.5}
-    );
-
-    sectionsRef.current.forEach((section) => {
-      if (section) observer.observe(section)
-    })
-
-    return () => {
-      sectionsRef.current.forEach((section) => {
-        if (section) observer.unobserve(section)
-      })
-    }
-  }, [])
-
-
-  return (
-    <>
-      <Navbar />  
-      <div className="scroll-container">
-        <div className="scroll-section" ref={(el) => (sectionsRef.current[0] = el)}>
-        <Header />
-        </div>
-        <div className="scroll-section skills" ref={(el) => (sectionsRef.current[1] = el)}>
-        <Skills />
-        </div>
-        <div className="scroll-section" ref={(el) => (sectionsRef.current[2] = el)}>
-        <CTA />
-       </div>
-        <div className="scroll-section" ref={(el) => (sectionsRef.current[3] = el)}>
-        <About />
-        </div>
-        <div className="scroll-section" ref={(el) => (sectionsRef.current[4] = el)}>
-        <Projects />
-        </div>
-        <div className="scroll-section" ref={(el) => (sectionsRef.current[5] = el)}>
-        <Contact />
-        </div>
-    </div>
-    </>
-  );
+  return route === "#/v2" ? <PortfolioV2 /> : <PortfolioCurrent />;
 }
 
 export default App;
