+++ src/pages/PortfolioV2.jsx
@@ -1,0 +1,26 @@
+import React from "react";
+
+function PortfolioV2() {
+  return (
+    <main
+      style={{
+        minHeight: "100vh",
+        background: "#111",
+        color: "#fff",
+        padding: "4rem 2rem",
+      }}
+    >
+      <section style={{ minHeight: "100vh" }}>
+        <h1>Portfolio V2</h1>
+        <p>New landing page in progress.</p>
+      </section>
+
+      <section style={{ minHeight: "100vh" }}>
+        <h2>Scene 2</h2>
+        <p>This is where the cinematic scroll experience will go.</p>
+      </section>
+    </main>
+  );
+}
+
+export default PortfolioV2;
